backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Maksim Zayakin <maksim.zaiakin@urfu.me>"]

[tool.poetry.dependencies]
python = "^3.8"
orjson = "^3.6.3"
odmantic = "^0.3.2"
fastapi = "^0.65.2"
uvicorn = "^0.13.2"
pysnmp = "^4.4.12"
loguru = "^0.5.3"
httpx = "^0.16.1"
python-multipart = "^0.0.5"
pymongo-migrate = "^0.11.0"
fastapi-mail = "^0.3.5"
celery = "^5.1.2"
redis = "^3.5.3"
flower = "^1.0.0"
python-snap7 = "^1.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
flake8 = "^3.8.4"
isort = "^5.6.4"
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
